requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
executing==1.2.0
filelock==3.12.1
fonttools==4.39.4
frozenlist==1.3.3
graphviz==0.20.1
grpcio==1.51.3
ibm-cloud-sdk-core==3.16.7
ibm-platform-services==0.36.0
idna==3.4
importlib-resources==5.12.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lark-parser==0.12.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
ntlm-auth==1.5.0
numpy==1.23.5
packaging==23.1
pandas==2.0.2
parse==1.19.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytket==1.15.0
pytket-qiskit==0.39.0
pytz==2023.3
PyYAML==6.0
qiskit==0.42.1
qiskit-aer==0.12.0
qiskit-ibm-runtime==0.9.3
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.23.3
qwasm==1.0.1
ray==2.5.0
requests==2.31.0
requests-ntlm==1.1.0
rustworkx==0.13.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
stevedore==5.1.0
stim==1.11.0
symengine==0.9.2
sympy==1.12
traitlets==5.9.0
types-pkg-resources==0.1.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
zipp==3.15.0
